完成表情管理接口

+++ app/enterprise/controller/Emoji.php
@@ -12,20 +12,18 @@ class Emoji extends BaseController
     // 表情列表
     public function index()
     {
-        $param = $this->request->param();
-        $listRows = $param['limit'] ?: 20;
-        $pageSize = $param['page'] ?: 1;
         $map=['status'=>1,'user_id'=>$this->uid,'type'=>2];
-        $list = EmojiModel::where($map)->field('id,name,src')->limit('update_time desc')->paginate(['list_rows'=>$listRows,'page'=>$pageSize]);
+        $list = EmojiModel::where($map)->field('id,name,src,file_id')->order('update_time desc')->select();
         $data=[];
         if($list){
-            $data=$list->toArray()['data'];
+            $data=$list->toArray();
             foreach ($data as $k => $v) {
                 $url=getFileUrl($v['src']);
                 $data[$k]['src'] =$url;
+                $data[$k]['title'] =$v['name'];
             }
         }
-        return success('', $data, $list->total(), $list->currentPage());
+        return success('', $data, count($data));
     }
 
     // 添加表情
@@ -40,7 +38,6 @@ class Emoji extends BaseController
         // 判断是否已经有了当前表情，有了就更新
         if($exist){
             EmojiModel::where(['id'=>$exist['id']])->update(['update_time'=>time()]);
-
         }else{
             $info=[
                 'user_id'=>$this->uid,
@@ -51,28 +48,49 @@ class Emoji extends BaseController
             ];
             EmojiModel::create($info);
         }
-        return success('');
+        return success(lang('system.addOk'));
     }
 
     // 删除表情
     public function del(){
-        $param = $this->request->param();
-        $id=$param['id'];
-        $emoji=EmojiModel::find($id);
-        if(!$emoji){
-            return warning(lang('system.exits'));
+        $ids = $this->request->param('ids',[]);
+        if(!is_array($ids) || $ids==[]){
+            return warning(lang('system.parameterError'));
         }
-        $res=EmojiModel::where(['id'=>$id])->delete();
-        if($res){
-            $exist=EmojiModel::where(['file_id'=>$emoji->file_id])->find();
-            // 如果文件没有引用了，就删除掉源文件
-            if(!$exist){
-                $disk=env('filesystem.driver','local');
-                $file=File::find($emoji->file_id);
-                Filesystem::disk($disk)->delete($file->src);
+        foreach($ids as $id){
+            $emoji=EmojiModel::where(['id'=>$id,'user_id'=>$this->uid])->find();
+            if(!$emoji){
+                continue;
+            }
+            $res=EmojiModel::where(['id'=>$id])->delete();
+            if($res){
+                $exist=EmojiModel::where(['file_id'=>$emoji['file_id']])->find();
+                // 如果文件没有引用了，就删除掉源文件
+                if(!$exist){
+                    $disk=env('filesystem.driver','local');
+                    $file=File::find($emoji['file_id']);
+                    Filesystem::disk($disk)->delete($file->src);
+                }
             }
         }
         return success(lang('system.delOk'));
+        
+    }
+
+    // 移动表情
+    public function move(){
+        $ids = $this->request->param('ids',[]);
+        if(!is_array($ids) || $ids==[]){
+            return warning(lang('system.parameterError'));
+        }
+        foreach($ids as $id){
+            $emoji=EmojiModel::where(['id'=>$id,'user_id'=>$this->uid])->find();
+            if(!$emoji){
+                continue;
+            }
+            EmojiModel::where(['id'=>$id])->update(['update_time'=>time()]);
+        }
+        return success(lang('system.success'));
         
     }
